Moved functions to appropriate files and adjusted sw route
