+++ src/utils/customerUtils.js
@@ -1,0 +1,72 @@
+
+import { openDB } from 'idb';
+
+// We define a customer to be the same if they have the same
+export const checkSameCustomer = (customerOne, customerTwo) => {
+    // Check rids if available, check those
+    if (customerOne.rid && customerTwo.rid && customerOne.rid === customerTwo.rid) {
+        return true;
+    }
+
+    const firstName = customerOne.name;
+    const secondName = customerTwo.name;
+    const firstMeterNumber = customerOne.meterNumber;
+    const secondMeterNumber = customerTwo.meterNumber;
+
+    return firstName && secondName && firstName === secondName && firstMeterNumber === secondMeterNumber;
+}
+
+export const addToOfflineCustomer = async (customer, fieldToAppend, objectToAdd) => {
+    let db, getTransaction, objectStore, requestCursor;
+    const requestKeyToValue = {};
+    try {
+        db = await openDB('workbox-background-sync');
+        getTransaction = db.transaction('requests');
+        objectStore = getTransaction.objectStore('requests');
+        requestCursor = await objectStore.openCursor();
+    } catch (err) {
+        console.log("Error when opening DB transaction. Exiting without modifying request.")
+        return;
+    }
+
+    console.log(customer, fieldToAppend, objectToAdd);
+
+    // In this transaction, iterate through all the requests and store them in the requestKeyValues
+    while (requestCursor) {
+        try {
+            const requestValue = requestCursor.value;
+            const requestKey = requestCursor.key;
+            console.log(requestKey, requestValue);
+
+            if (requestValue && requestValue.storableRequest && requestValue.storableRequest.requestInit) {
+                requestKeyToValue[requestKey] = requestValue;
+            }
+        } catch (error) {
+            console.log('Error when iterating through cursor on key ', requestCursor.key);
+        }
+        requestCursor = await requestCursor.continue();
+    }
+    await getTransaction.done;
+
+    // Decode the requestBlobs and replace if appropriate
+    for (const [requestKeyString, requestValue] of Object.entries(requestKeyToValue)) {
+        console.log("Request value: ", requestValue);
+        const requestBodyBlob = requestValue.storableRequest.requestInit.body;
+        const requestBody = await requestBodyBlob.text();
+        const createCustomerJSON = JSON.parse(requestBody);
+
+        if (checkSameCustomer(customer, createCustomerJSON)) {
+            const modifiedRequest = { ...requestValue };
+            createCustomerJSON[fieldToAppend].push(objectToAdd);
+            const modifiedCustomerJSON = JSON.stringify(createCustomerJSON);
+            modifiedRequest.storableRequest.requestInit.body = new Blob([modifiedCustomerJSON], { type: 'application/json' });
+
+            const requestKey = isNaN(requestKeyString) ? requestKeyString : parseInt(requestKeyString);
+
+            console.log("Modified request: ", modifiedRequest);
+            await db.put('requests', modifiedRequest, requestKey);
+        } else {
+            console.log("Not same customer!");
+        }
+    }
+}
