+++ src/utils/new-schema/request.js
@@ -23,6 +23,7 @@ import {
   getRecordById,
   deleteRecord,
 } from './airtable';
+import { addToOfflineCustomer } from '../customerUtils';
 
 /*
  ******* CREATE RECORDS *******
@@ -73,11 +74,12 @@ export const createManyTariffPlans = async (records) => {
   return Promise.all(createPromises);
 };
 
-// We use a special, non-schema-generated createCustomer that hits a special endpoint
-// because we require additional logic to handle offline functionality
-const createCustomer = async (customer) => {
+// NONGENERATED: We use a special, non-schema-generated createCustomer 
+// that hits a special endpoint because we require additional logic to 
+// handle offline functionality
+export const createCustomer = async (customer) => {
   try {
-    const resp = await fetch('http://127.0.0.1:4000/customer/create', {
+    const resp = await fetch('http://127.0.0.1:4000/customers/create', {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
@@ -107,9 +109,30 @@ export const createManyCustomerUpdates = async (records) => {
   return Promise.all(createPromises);
 };
 
-export const createMeterReadingsandInvoice = async (record) => {
-  return createRecord(Tables.MeterReadingsandInvoices, record);
-};
+// NONGENERATED: Create a meter reading for a customer
+export const createMeterReadingsandInvoice = async (meterReading, customer) => {
+  // If customer does not exist, we want to search the requests objectStore
+  // to add the current meter reading to the customer request being POST'ed
+  if (!customer.rid) {
+    addToOfflineCustomer(customer, 'meterReadings', meterReading)
+  } else {
+    // Customer has an rid so it is in the airtable.
+    // Make a standard request to create a meter reading / invoice.
+    try {
+      meterReading.customerId = customer.rid;
+      const resp = await fetch('http://127.0.0.1:4000/meter-readings-and-invoices/create', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(meterReading)
+      })
+      console.log('Response for meter reading: ', resp);
+    } catch (err) {
+      console.log('Error with create meter reading request: ', err);
+    }
+  }
+}
 
 export const createManyMeterReadingsandInvoices = async (records) => {
   const createPromises = [];
@@ -122,9 +145,30 @@ export const createManyMeterReadingsandInvoices = async (records) => {
   return Promise.all(createPromises);
 };
 
-export const createPayment = async (record) => {
-  return createRecord(Tables.Payments, record);
-};
+// NONGENERATED: Create a payment for a customer
+export const createPayment = async (payment, customer) => {
+  // If customer does not exist, we want to search the requests objectStore
+  // to add the current meter reading to the customer request being POST'ed
+  if (!customer.rid) {
+    addToOfflineCustomer(customer, 'payments', payment);
+  } else {
+    // Customer has an rid so it is in the airtable.
+    // Make a standard request to create a payment.
+    try {
+      payment.customerId = customer.rid;
+      const resp = await fetch('http://127.0.0.1:4000/payments/create', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(payment)
+      })
+      console.log('Response for payment: ', resp);
+    } catch (err) {
+      console.log('Error with create payment request: ', err)
+    }
+  }
+}
 
 export const createManyPayments = async (records) => {
   const createPromises = [];
